+++ basis/common-router/src/main/java/com/softgroup/common/router/api/factories/HandlerFactory.java
@@ -7,5 +7,5 @@ import com.softgroup.common.router.api.Handler;
  * Created by Виктор on 03.03.2017.
  */
 public interface HandlerFactory<T extends Handler> {
-    Handler getHandler(Request<?> msg);
+    Handler getHandler(String routeKey);
 }
